# Return iterator of InternalRow directly from library