--- a/src/main/java/com/github/sadikovi/netflowlib/Strategies.java
+++ b/src/main/java/com/github/sadikovi/netflowlib/Strategies.java
@@ -16,7 +16,6 @@
 
 package com.github.sadikovi.netflowlib;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.HashMap;