--- a/src/main/java/com/github/sadikovi/netflowlib/record/PredicateRecordMaterializer.java
+++ b/src/main/java/com/github/sadikovi/netflowlib/record/PredicateRecordMaterializer.java
@@ -96,12 +96,13 @@ public final class PredicateRecordMaterializer extends RecordMaterializer implem
       int i = 0, offset = 0, len = 0;
       while (i < numColumns) {
         len = columns[i].getUnsignedBytes();
-        offsets[i] = offset;
         System.arraycopy(buffer.array(), buffer.arrayOffset() + columns[i].getColumnOffset(),
           data, offset, len);
+        offsets[i] = offset;
         offset += len;
+        ++i;
       }
-      return new ByteBufRow(offsets, data);
+      return new ByteBufRow(offsets, data, buffer.order());
     }
   }