… src/main/scala/com/github/sadikovi/spark/netflow/version7/DefaultProvider.scala
… src/test/scala/com/github/sadikovi/spark/netflow/sources/testproviders.scala