This PR adds row iterator to the netflow reader to return `Iterator<InternalRow>` and make it compatible with Spark SQL. Conversion is implemented as copy into new internal row with updated fields. Benchmarks show that this change is slower than current master, but changes exposed couple of improvements that can be made.

Master (2.0.2):
```
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
NetFlow full scan:                  Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------
Scan, stringify = F                      9664 / 9810        103.5      966383.4       1.0X
Scan, stringify = T                    20641 / 20772         48.4     2064072.8       0.5X

Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
NetFlow predicate scan:             Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------
Predicate pushdown = F, high           23778 / 24084         42.1     2377846.1       1.0X
Predicate pushdown = T, high           23411 / 23569         42.7     2341062.5       1.0X
Predicate pushdown = F, low            13737 / 13833         72.8     1373701.3       1.7X
Predicate pushdown = T, low              3169 / 3210        315.6      316861.1       7.5X
```

This patch:
```
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
NetFlow full scan:                  Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------
Scan, stringify = F                    12426 / 12475         80.5     1242558.0       1.0X
Scan, stringify = T                    35894 / 36084         27.9     3589417.5       0.3X

Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
NetFlow predicate scan:             Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
-------------------------------------------------------------------------------------------
Predicate pushdown = F, high           40144 / 40220         24.9     4014354.2       1.0X
Predicate pushdown = T, high           40680 / 41479         24.6     4068008.7       1.0X
Predicate pushdown = F, low            28519 / 28655         35.1     2851922.8       1.4X
Predicate pushdown = T, low              3211 / 3302        311.4      321085.1      12.5X
```